fix: align coordinator playground with shared prompt template
