+++ packages/agents/src/plugins/plugin-user-core/src/index.ts
@@ -96,7 +96,6 @@ export {
   relayToAgentAction,
 } from './actions';
 export {
-  COORDINATOR_CONTEXT_TEXT,
   coordinatorActionStateProvider,
   coordinatorActionsProvider,
   coordinatorAgentActivityProvider,
@@ -104,7 +103,6 @@ export {
   coordinatorDispatchHistoryProvider,
   coordinatorRecentMessagesProvider,
   coordinatorTeamMembersProvider,
-  formatActionsWithParams,
 } from './providers';
 export * from './types';
